refactor: 优化目录结构
