+++ src/components/section-manager/editorTypeConstants.ts
